RDKCOM-3435: RDKBCMF-296 Utopia build failure in RPi4 Broadband 64bit
Reason for change: build errors in utopia while building RPi4 Broadband 64bit:
| Utopia/source/service_dhcp/service_dhcp_server.c:221:58: error: format '%d' expects argument of type 'int', but argument 5 has type 'long unsigned int' [-Werror=format=]
|   221 |             fprintf(stdout, DNSMASQ getRFC_Value %s %s %d\n,status,l_DnsStrictOrderStatus,sizeof(l_DnsStrictOrderStatus));
|       |                                                         ~^                                  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
|       |                                                          |                                  |
|       |                                                          int                                long unsigned int
|       |                                                         %ld
| Utopia/source/service_dhcp/service_dhcp_server.c:225:78: error: format '%d' expects argument of type 'int', but argument 5 has type 'long unsigned int' [-Werror=format=]
|   225 |             fprintf(stdout, FAILURE: DNSMASQ getRFC_Value syscfg_get %s %s %d\n,status,l_DnsStrictOrderStatus,sizeof(l_DnsStrictOrderStatus));
|       |                                                                             ~^                                  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
|       |                                                                              |                                  |
|       |                                                                              int                                long unsigned int
|       |                                                                             %ld

Test Procedure: Broadband 64bit build OK

Risks: None.

(cherry picked from commit 05cabcdf9d9c90ac33164e075d4058b6fb032bc5)

+++ source/service_dhcp/service_dhcp_server.c
@@ -219,11 +219,11 @@ void getRFC_Value(const char* dnsOption)
         result = strcmp (status,l_DnsStrictOrderStatus);
         if (result == 0){
             strncpy((char *)dnsOption,dnsSet, strlen(dnsSet));
-            fprintf(stdout, "DNSMASQ getRFC_Value %s %s %d\n",status,l_DnsStrictOrderStatus,sizeof(l_DnsStrictOrderStatus));
+            fprintf(stdout, "DNSMASQ getRFC_Value %s %s %ld\n",status,l_DnsStrictOrderStatus, (long unsigned int)sizeof(l_DnsStrictOrderStatus));
             fprintf(g_fArmConsoleLog, "Starting dnsmasq with additional dns strict order option: %s\n",l_DnsStrictOrderStatus);
         }
         else{
-            fprintf(stdout, "FAILURE: DNSMASQ getRFC_Value syscfg_get %s %s %d\n",status,l_DnsStrictOrderStatus,sizeof(l_DnsStrictOrderStatus)); 
+            fprintf(stdout, "FAILURE: DNSMASQ getRFC_Value syscfg_get %s %s %ld\n",status,l_DnsStrictOrderStatus, (long unsigned int)sizeof(l_DnsStrictOrderStatus)); 
             fprintf(g_fArmConsoleLog, "RFC DNSTRICT ORDER is not defined or Enabled %s\n", l_DnsStrictOrderStatus);
         }
 }
